backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
deepface>=0.0.93
chromadb>=0.5.0
python-multipart>=0.0.9
requests>=2.31.0
numpy>=1.26.0
tf-keras>=2.16.0
python-dotenv>=1.0.0

pillow>=10.0.0
